refactor: assemble the overall structure
