+++ src/app/app.module.ts
@@ -4,7 +4,7 @@ import { BrowserModule } from '@angular/platform-browser';
 
 import { AppRoutingModule } from 'src/app/app-routing.module';
 import { AppComponent } from 'src/app/app.component';
-import { SharedRootModule } from 'src/app/shared/modules/shared-root/shared-root.module';
+import { SharedRootModule } from 'src/app/shared/shared-root/shared-root.module';
 
 @NgModule({
   declarations: [AppComponent],
