+++ src/app/core/components/header/account/account.component.scss
