+++ src/app/core/components/header/search/search.component.html
@@ -1,4 +1,4 @@
- <img class="search_logo" src="./assets/Images/youtube.png" alt="youtube logo">
+ <img class="search_logo" src="../../../../../assets/Images/youtube.png" alt="youtube logo">
 
   <mat-form-field style="width: 30vw;padding-top: 1.5rem">
     <mat-label>{{btnSearchTitle|titlecase}}</mat-label>
