+++ src/app/core/components/header/search/search.component.ts
