+++ src/app/shared/shared-root/components/shared-root/shared-root.component.ts
@@ -1,6 +1,6 @@
 import { Component } from '@angular/core';
-import { mockEntities } from '../../../../../mocks/mock';
-import { ResponseItem } from '../../../../models/response-item/response-item.model';
+import { mockEntities } from '../../../../mocks/mock';
+import { ResponseItem } from '../../../models/response-item/response-item.model';
 
 @Component({
   selector: 'app-shared-root',
